Refactor: Token Picker modal centralized in common component
